Move secretsmanager to aws dir
