+++ internal/pkg/aws/secretsmanager/secretsmanager_test.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/store/secretsmanager/mocks"
+	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/aws/secretsmanager/mocks"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/golang/mock/gomock"
